Split extensions tests per header
