+++ tests/src/extensions/gtest_tests.cpp
@@ -1,0 +1,24 @@
+#include <sstream>
+
+#include <gtest/gtest.h>
+#include <stronk/extensions/gtest.h>
+#include <stronk/stronk.h>
+
+namespace twig
+{
+
+struct a_can_gtest_print_type : stronk<a_can_gtest_print_type, int, can_gtest_print>
+{
+    using stronk::stronk;
+};
+
+TEST(can_gtest_print, streaming_to_ostream_prints_the_value)  // NOLINT
+{
+    auto val = a_can_gtest_print_type {5};
+    auto sstream = std::stringstream();
+    PrintTo(val, &sstream);
+
+    EXPECT_EQ(sstream.str(), "5");
+}
+
+}  // namespace twig
